requirements.txt
streamlit
numpy
Pillow
scikit-learn
kneed
scikit-image